consolidate to a single package to make it easier to deploy
