+++ pyproject.toml
@@ -1,24 +1,23 @@
 [project]
-name = "stockdice-app"
+name = "stockdice"
 version = "0.1.0"
-description = "Add your description here"
+description = "A random stock picker working with FMP API"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
-    "stockdice",
+    "flask>=3.1.1",
+    "google-cloud-storage>=3.2.0",
+    "httpx>=0.28.1",
+    "numpy>=2.3.2",
+    "polars>=1.32.0",
+    "pyarrow>=21.0.0",
+    "toml>=0.10.2",
 ]
 
-[tool.uv.workspace]
-members = [
-    "shared-library",
-    "webapp",
-]
+[build-system]
+requires = ["setuptools >= 61.0"]
+build-backend = "setuptools.build_meta"
 
-[tool.uv.sources]
-stockdice = { workspace = true }
-
-[dependency-groups]
-dev = [
-    "ipython>=9.4.0",
-    "ruff>=0.12.7",
-]
+[tool.setuptools.packages.find]
+where = ["./"]
+include = ["stockdice"]
