--- shared-library/main.py
@@ -1,6 +1,0 @@
-def main():
-    print("Hello from shared-library!")
-
-
-if __name__ == "__main__":
-    main()
