+++ specs/2025-08-04-initial-web-app.md
@@ -1,6 +1,6 @@
-# Title of the Specification
+# Roll the stock dice
 
-*Provide a brief overview of the feature or bug.*
+
 
 ## Background
