+++ stockdice/dice.py
