+++ stockdice/income.py
