+++ stockdice/stocklist.py
