+++ stockdice/timeutils.py
