+++ tests/unit/test_trading_hours.py
