--- webapp/README.md
@@ -1,1 +1,0 @@
-Web app powered by the stockdice package.
