--- webapp/pyproject.toml
@@ -1,9 +1,0 @@
-[project]
-name = "stockdiceapp"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.12"
-dependencies = [
-    "flask>=3.1.1",
-]
